Add user info to Nav bar.
